+++ src/components/ShareCanvasLinkDialog.jsx
@@ -1,34 +1,18 @@
-import Box from "@material-ui/core/Box";
-import Button from "@material-ui/core/Button";
-import ButtonGroup from "@material-ui/core/ButtonGroup";
-import Dialog from "@material-ui/core/Dialog";
-import DialogActions from "@material-ui/core/DialogActions";
-import DialogTitle from "@material-ui/core/DialogTitle";
-import makeStyles from "@material-ui/core/styles/makeStyles";
-import TextField from "@material-ui/core/TextField";
-import Typography from "@material-ui/core/Typography";
-import Alert from "@material-ui/lab/Alert";
-import ns from "mirador/dist/es/src/config/css-ns";
-import ScrollIndicatedDialogContent from "mirador/dist/es/src/containers/ScrollIndicatedDialogContent";
+import Alert from "@mui/material/Alert";
+import Box from "@mui/material/Box";
+import Button from "@mui/material/Button";
+import ButtonGroup from "@mui/material/ButtonGroup";
+import Dialog from "@mui/material/Dialog";
+import DialogActions from "@mui/material/DialogActions";
+import DialogTitle from "@mui/material/DialogTitle";
+import TextField from "@mui/material/TextField";
+import { ScrollIndicatedDialogContent, useTranslation } from "mirador";
 import PropTypes from "prop-types";
-import React, { useState } from "react";
+import { useState } from "react";
 
 import CopyToClipboard from "./dialog/CopyToClipboard";
 import RightsInformation from "./dialog/RightsInformation";
 import ShareButton from "./dialog/ShareButton";
-
-const useStyles = makeStyles((theme) => ({
-  actions: {
-    justifyContent: "space-between",
-    flexWrap: "wrap",
-  },
-  actionButtons: {
-    flexWrap: "wrap",
-  },
-  alert: {
-    marginBottom: theme.spacing(1),
-  },
-}));
 
 const supportsClipboard = "clipboard" in navigator;
 
@@ -40,13 +24,12 @@ const ShareCanvasLinkDialog = ({
   visibleCanvases,
   label,
   rights,
-  t,
   updateConfig,
   windowViewType,
 }) => {
   const { dialogOpen, enabled, showRightsInformation, getCanvasLink } = config;
   const [copiedToClipboard, setCopiedToClipboard] = useState(false);
-  const { actions, actionButtons, alert } = useStyles();
+  const { t } = useTranslation();
 
   if (!enabled || !dialogOpen || visibleCanvases.length === 0) {
     return null;
@@ -67,22 +50,22 @@ const ShareCanvasLinkDialog = ({
 
   return (
     <Dialog
-      container={document.querySelector(`#${containerId} .${ns("viewer")}`)}
+      container={document.querySelector(`#${containerId} .mirador-viewer`)}
       fullWidth
       maxWidth="sm"
       scroll="paper"
       open={dialogOpen}
       onClose={closeDialog}
     >
-      <DialogTitle disableTypography>
-        <Typography variant="h4">
-          <Box fontWeight="fontWeightBold">{t("canvasLink.shareLink")}</Box>
-        </Typography>
+      <DialogTitle>
+        <Box fontWeight="fontWeightBold" component="span">
+          {t("canvasLink.shareLink")}
+        </Box>
       </DialogTitle>
       <ScrollIndicatedDialogContent dividers>
         {copiedToClipboard && (
           <Alert
-            className={alert}
+            sx={{ mb: 1 }}
             closeText={t("canvasLink.close")}
             onClose={() => setCopiedToClipboard(false)}
             severity="success"
@@ -101,7 +84,6 @@ const ShareCanvasLinkDialog = ({
                   setTimeout(() => setCopiedToClipboard(false), 3000);
                 }}
                 supported={supportsClipboard}
-                t={t}
               />
             ),
             readOnly: true,
@@ -110,10 +92,19 @@ const ShareCanvasLinkDialog = ({
           value={canvasLink}
           variant="outlined"
         />
-        {showRightsInformation && <RightsInformation t={t} rights={rights} />}
+        {showRightsInformation && <RightsInformation rights={rights} />}
       </ScrollIndicatedDialogContent>
-      <DialogActions className={actions}>
-        <ButtonGroup className={actionButtons}>
+      <DialogActions
+        sx={{
+          justifyContent: "space-between",
+          flexWrap: "wrap",
+        }}
+      >
+        <ButtonGroup
+          sx={{
+            flexWrap: "wrap",
+          }}
+        >
           {["envelope", "facebook", "pinterest", "x", "whatsapp"].map((p) => (
             <ShareButton
               canvasLink={canvasLink}
@@ -157,7 +148,6 @@ ShareCanvasLinkDialog.propTypes = {
   label: PropTypes.string,
   manifestId: PropTypes.string.isRequired,
   rights: PropTypes.arrayOf(PropTypes.string),
-  t: PropTypes.func.isRequired,
   updateConfig: PropTypes.func.isRequired,
   visibleCanvases: PropTypes.arrayOf(
     PropTypes.shape({