+++ src/components/ShareControl.jsx
@@ -1,16 +1,10 @@
-import ShareIcon from "@material-ui/icons/Share";
-import { MiradorMenuButton } from "mirador/dist/es/src/components/MiradorMenuButton";
+import ShareIcon from "@mui/icons-material/Share";
+import { MiradorMenuButton, useTranslation } from "mirador";
 import PropTypes from "prop-types";
-import React from "react";
 
-const ShareControl = ({
-  containerId,
-  config,
-  t,
-  updateConfig,
-  windowViewType,
-}) => {
-  const { dialogOpen, enabled, singleCanvasOnly } = config;
+const ShareControl = ({ config, updateConfig, windowViewType }) => {
+  const { dialogOpen, enabled, singleCanvasOnly } = config || {};
+  const { t } = useTranslation();
   if (
     !enabled ||
     // Only show in single canvas view if configured
@@ -24,7 +18,6 @@ const ShareControl = ({
     <MiradorMenuButton
       aria-expanded={dialogOpen}
       aria-label={t("canvasLink.shareLink")}
-      containerId={containerId}
       onClick={() =>
         updateConfig({
           ...config,
@@ -43,8 +36,6 @@ ShareControl.propTypes = {
     enabled: PropTypes.bool.isRequired,
     singleCanvasOnly: PropTypes.bool.isRequired,
   }).isRequired,
-  containerId: PropTypes.string.isRequired,
-  t: PropTypes.func.isRequired,
   updateConfig: PropTypes.func.isRequired,
   windowViewType: PropTypes.string.isRequired,
 };