+++ src/components/dialog/RightsInformation.jsx
@@ -1,23 +1,16 @@
-import Link from "@material-ui/core/Link";
-import makeStyles from "@material-ui/core/styles/makeStyles";
-import Alert from "@material-ui/lab/Alert";
+import Alert from "@mui/material/Alert";
+import Link from "@mui/material/Link";
+import { useTranslation } from "mirador";
 import PropTypes from "prop-types";
-import React from "react";
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    marginTop: theme.spacing(2),
-  },
-}));
 
 /** Renders the rights information defined in the used manifest */
-const RightsInformation = ({ rights, t }) => {
-  const { root } = useStyles();
+const RightsInformation = ({ rights }) => {
+  const { t } = useTranslation();
   if (!rights.length) {
     return null;
   }
   return (
-    <Alert className={root} severity="warning">
+    <Alert sx={{ mt: 2 }} severity="warning">
       <span>{t("canvasLink.noteRights", { count: rights.length })}: </span>
       {rights.length === 1 ? (
         <Link href={rights[0]} rel="noopener" target="_blank">
@@ -26,7 +19,7 @@ const RightsInformation = ({ rights, t }) => {
       ) : (
         <ul>
           {rights.map((link) => (
-            <li>
+            <li key={link}>
               <Link href={link} rel="noopener" target="_blank">
                 {link}
               </Link>
@@ -40,7 +33,6 @@ const RightsInformation = ({ rights, t }) => {
 
 RightsInformation.propTypes = {
   rights: PropTypes.arrayOf(PropTypes.string).isRequired,
-  t: PropTypes.func.isRequired,
 };
 
 export default RightsInformation;