+++ src/components/icons/XIcon.jsx
@@ -1,5 +1,4 @@
-import { SvgIcon } from "@material-ui/core";
-import React from "react";
+import { SvgIcon } from "@mui/material";
 
 /* FIXME: should not be needed anymore when upgrading to MUI 5, see https://github.com/mui/material-ui/pull/38811 */
 const XIcon = () => (